[Wave] Remove Wave-related code and docs (#1128)
As announced in https://github.com/iree-org/iree-turbine/issues/1055,
Wave project has moved into a new repository. This is amongst a series
of PR that removes Wave related code and document from `iree-turbine`
repo. In a previous patch tests were removed
(https://github.com/iree-org/iree-turbine/pull/1110) and this PR removes
all documents and most of the Wave code with a couple of exceptions.

In a follow-up PR, `wave_ops.py` and `wave_types.py` need to be removed.
Currently there are methods from these files that are used by tk kernels
(in tests) which needs to be reconciled.

---------

Signed-off-by: nithinsubbiah <nithinsubbiah@gmail.com>