Make iree.trace_tensor a side-effecting op
Instead of abusing the in-place mechanism make the side-effecting.

This relies on a nonpublic functionality, so it is likely that it will break in
future torch versions.
We really want to make the op side-effecting because this is really the correct semantics.
If we don't have this we need to declare the op as mutating the tensor argument, which
causes problems when trying to trace frozen tensors.
It will also cause problems when for example we trace a tensor that does not affect
the result of the export. Then the trace op will be removed as dead code.
E.g.
```
y = x.clone()
trace_tensor(y)
```

Signed-off-by: Boian Petkantchin <boian.petkantchin@amd.com>