[boo] add explicit casts for layer norm intermediates (#1107)
Torch choses to store intermediate values passes from the forward to the
backward pass using different dtypes based on the device: on CPU, the
same dtype as the input is used even if the computation uses double
bitwidth, on GPU, the double-width dtype is used instead. Guard against
this by allowing the user to explicitly specify the dtype to use for
intermediates in the layer norm signature. Note that exporting from
torch will still introduce truncations, but these are expected to cancel
out during compilation.

Signed-off-by: Alex Zinenko <git@ozinenko.com>