[boo] use torch compile backend in driver
Go through `torch.compile` in the boo driver instead of calling the
kernel directly. This allows to collect more realistic execution time
statistics that are close to what the final user will see after
integration. The direct kernel invocation is kept as a reference option
to evaluate the overhead.

Tweak the profiler configuration to ignore initialization and cleanup
steps so as to avoid skewing aggregate statistics.

Signed-off-by: Alex Zinenko <git@ozinenko.com>