[BOO] Ensure driver calls synchronize on all devices (#1146)
Previously, when running a test on multiple GPUs, it was only
synchronizing on the default device after a run, and was reporting fewer
iterations than expected. This change simply changes the
`torch.cuda.synchronize` call to instead synchronize on all devices used
for the test.

A single unrelated change is included to add a shorter alias for the
`--verbose` option.

Signed-off-by: zjgarvey <zjgarvey@gmail.com>