[boo] add experimental torch.compile backend variant, make layernorm go through it (#1135)
Add a second, experimental BOO backend for the purposes of performance
comparisons. It is allowed to have more fusion specs, including
unprofitable ones and is not going to be on by default.

Add a pseudo-fusion spec that intercepts layer norm operations, but
doesn't actually do fusion since we don't know at this point what do we
want to fuse with. Known-good operations, without permutations, are
accepted by the main BOO backend. The remaining operations are only
accepted by the experimental backend.

---------

Signed-off-by: Alex Zinenko <git@ozinenko.com>