[BOO] Make new sample inputs during `get_launchable` for graph ops (#1159)
This is to protect against certain metadata from getting attached to the
subgraph inputs.

I haven't been able to reproduce this issue outside of running a real
model yet, so writing a test is proving difficult.

The issue happens because one of the inputs for the `torch.export` call
ends up having some metadata saying it's a parameter of the module, but
`LayoutManagedModuleForAOTExport` doesn't have any params. It throws a
`KeyError` when trying to rename the missing parameter.

This change will simply sterilize any sample inputs for the export
process. I'll try to isolate a test for this in a follow up PR.

---------

Signed-off-by: zjgarvey <zjgarvey@gmail.com>