[boo driver] Add support for tsv command files (#1182)
This adds support for consuming tab-separated commands on the
command-line and from `.tsv` command files.

For example,
```
$ iree-boo-driver "aten::mm	((3840, 4352), (4352, 3840))	('c10::BFloat16', 'c10::BFloat16')	((1, 3840), (3840, 1))	('', '')"
```
and
```
$ iree-boo-driver --commands-file commands.tsv
```
```tsv
# commands.tsv
aten::mm	((3840, 4352), (4352, 3840))	('c10::BFloat16', 'c10::BFloat16')	((1, 3840), (3840, 1))	('', '')
```
are now equivalent to
```
$ iree-boo-driver aten::mm "((3840, 4352), (4352, 3840))" "('c10::BFloat16', 'c10::BFloat16')" "((1, 3840), (3840, 1))" "('', '')"
```

This is a convenience feature for working with excel sheets that have
driver arguments split into columns. Copying from these results in
tab-separated values.