[boo driver] Add aten.addmm parser (#1193)
This adds support for `aten.addmm` in the driver. The main addition here
is support for concrete arguments, which is used for the `beta` and
`alpha` arguments. `addmm`'s signature:
```
addmm(Tensor self, Tensor mat1, Tensor mat2, *, Scalar beta=1, Scalar alpha=1) -> Tensor
```
and example command:
```
aten::addmm "[[512], [24576, 512], [512, 512], [], []]" "['c10::BFloat16', 'c10::BFloat16', 'c10::BFloat16', 'Scalar', 'Scalar']" "[[1], [512, 1], [1, 512], [], []]" "['', '', '', '1', '1']"
```

The scalar arguments are only accepted as keyword args, however the
pytorch profile output encodes the arguments as a flat list. The torch
op schema is used to find the argument name for each concrete arg so
they can be passed in as keyword arguments when the torch op is invoked
(this assumes that the order that the torch profile uses will match the
order of arguments in the torch op schema, which from my experiments
seems to be the case).