[boo] Use channels-last layout for custom conv op with 3D convs (#1211)
The missing check here was causing 3D NDHWC convs to have IR emitted
with a separate transpose on the output, rather than the conv directly
computing a channels-last output layout.

With this change, performance for 3D convs is the same on both
`iree_boo_legacy` and `iree_boo_experimental` backends.