[BOO] Rework graph ops to avoid re-exporting layout managed modules (#1170)
This PR modifies the subgraph `GraphModule` to handle layouts instead of
creating a new `nn.Module` and re-exporting.

Since we aren't using `aot.export`, decompositions need to be applied
directly to the graph. Decomposing the subgraph aligns the ir output
with `aot.export` and has the added benefit of automatically
canonicalizing the graph module. Therefore, we will likely have far
fewer cache duplicates based on things like variable names or stack
traces.

Eventually, it would be rather convenient to combine decompositions and
replacements, since these are roughly equivalent.

There are some blockers to unifying decompositions and replacements.

1. We cannot easily infer the original node's output stride, and
performing a decomposition that results in a different layout is
difficult to propagate correctly in the rest of the model. E.g.,
replacing one conv might affect the input strides of the subsequent
conv, and we would need to re-prop before each subgraph
extraction/replacement. This seems rather untenable unless we can
re-work the replacement strategy into something which is also
propagating metadata at the same time.
2. I do not know how to selectively decompose (e.g., get a match failure
and *not* decompose).

---------

Signed-off-by: zjgarvey <zjgarvey@gmail.com>