[runtime][rocm] ensure we do not sync streams during dlpack handoff (#1230)
Recently, the meaning of `stream=None` has been changed to mean that we
are synchronizing with the default stream (see
https://github.com/pytorch/pytorch/commit/1d526fe78fcb8e6185f25d4b4340599b35f6a6d9
).

We could pass `t.__dlpack__(stream=-1)`, but recently added checks in
this method will actually error out if the `torch.cuda.current_device`
doesn't match that of `t` (this is likely to ensure the "default" stream
is appropriate, but we aren't doing any stream synchronization here).
Also, things like checking `t.layout == torch.strided_layout` are
automatic from our contiguous check. Therefore, I'm opting to bypass
`Tensor.__dlpack__` entirely, as all checks in this function are
completely redundant for us.

This change results in a reduction of a whopping 25us of CPU overhead
*per-pointer* on recent (nightly) pytorch versions.

---------

Signed-off-by: zjgarvey <zjgarvey@gmail.com>