[BOO] Enable in-place invocations for graph ops (#1220)
This change modifies a graph op's `GraphModule` to introduce output init
tensor placeholders and copy ops from the outputs to to these
placeholders. Additionally, we construct an `ExportGraphSignature` to
encode these input mutations, and directly assemble a minimal
`ExportedProgram` to pass to the `FxImporter`.

Although we could modify the graph module more simply and re-export with
torch.export, this became an untenable approach, since this export uses
gpu memory, and has issues with the pre-existing fake mode from
`torch.compile`.

Another approach could be manually modifying the exported MLIR, but this
would require translating quite a bit of information back and forth
between MLIR and pytorch.

---------

Signed-off-by: zjgarvey <zjgarvey@gmail.com>